requirements.txt
torch
torchvision
transformers
opencv-python
Pillow
numpy
tqdm
